Added VGA memory display
Switches 0 and 1 now control what is displayed over VGA. if both are 0 it is the complete contents of the register file. If switch 0 is high, it is locations 30-35 of memory and if switch 1 is high it is locations 36-41 of memory. Demonstrated that memory values are change in mode I though likely not correctly (or not displaying correctly)